+++ newclient.cpp
@@ -135,7 +135,7 @@ void NewClient::CenterForm()
     setFixedSize(geometry().width(), geometry().height());
     QRect desktopRect = QApplication::desktop()->availableGeometry(this);
     QPoint center = desktopRect.center();
-    move(center.x()-width()*0.5, center.y()-height()*0.5);
+    move(center.x()-static_cast<int>(width()*0.5), center.y()-static_cast<int>(height()*0.5) );
 }
 
 
@@ -158,7 +158,7 @@ void NewClient::RecordCarToClient()
 
 bool NewClient::CheckField(QString SelectedString)
 {
-    if(( (int)SelectedString.size() < 2 ) || SelectedString.isEmpty())
+    if(( static_cast<int>(SelectedString.size()) < 2 ) || SelectedString.isEmpty())
     {
         qDebug() << "NewClient::CheckField  FAIL SelectedString " << SelectedString;
         return false;
